+++ R/export_cdisc.R
@@ -232,17 +232,6 @@ export_cdisc <- function(res_nca, grouping_vars = character(0)) {
       } else {
         NA_character_
       },
-      ANL01FL = {
-        flag <- if ("ANL01FL" %in% names(.)) {
-          ANL01FL
-        } else {
-          rep("Y", nrow(.))
-        }
-        if ("is.excluded.hl" %in% names(.)) {
-          flag <- ifelse(is.excluded.hl, NA_character_, flag)
-        }
-        flag
-      },
       PKSUM1F = {
         flag <- if ("PKSUM1F" %in% names(.)) {
           PKSUM1F