+++ R/zzz.R
@@ -7,7 +7,6 @@
     ":=",
     "ADOSEDUR",
     "AFRLT",
-    "ANL01FL",
     "ARRLT",
     "ATPT",
     "ATPTN",